Change size type from int to double

+++ src/Envelope.cpp
@@ -6,22 +6,22 @@ Envelope::Envelope()
 	m_short_side_size = 1;
 }
 
-Envelope::Envelope(int size_1, int size_2)
+Envelope::Envelope(double size_1, double size_2)
 {
 	set_side_sizes(size_1, size_2);
 }
 
-int Envelope::get_long_side_size()
+double Envelope::get_long_side_size()
 {
 	return m_long_side_size;
 }
 
-int Envelope::get_short_side_size()
+double Envelope::get_short_side_size()
 {
 	return m_short_side_size;
 }
 
-void Envelope::set_side_sizes(int size_1, int size_2)
+void Envelope::set_side_sizes(double size_1, double size_2)
 {
 	if (!IsSizeValid(size_1))
 		throw InvalidEnvelopeSizeException("Invalid envelope size value", size_1);
@@ -41,7 +41,7 @@ void Envelope::set_side_sizes(int size_1, int size_2)
 	}
 }
 
-bool Envelope::IsSizeValid(int envelope_size)
+bool Envelope::IsSizeValid(double envelope_size)
 {
 	if (envelope_size <= 0)
 		return false;
