+++ src/Envelope.hpp
@@ -6,14 +6,14 @@ class Envelope
 {
 public:
 	Envelope();
-	Envelope(int side_1, int side_2);
-	int get_long_side_size();
-	int get_short_side_size();
-	void set_side_sizes(int size_1, int size_2);
+	Envelope(double side_1, double side_2);
+	double get_long_side_size();
+	double get_short_side_size();
+	void set_side_sizes(double size_1, double size_2);
 
 private:
-	int m_long_side_size;
-	int m_short_side_size;
-	bool IsSizeValid(int envelope_size);
+	double m_long_side_size;
+	double m_short_side_size;
+	bool IsSizeValid(double envelope_size);
 };
